/ Some more cleanup (moving code into the relevant classes).
